requirements.txt
langchain
langchain-community
pinecone
pinecone-plugin-assistant
python-dotenv
streamlit
sentence-transformers
google.api

llama-index-embeddings-openai
langchain_pinecone
langchain_openai